strategy: gzip + base64 encode the manifest
Manifests in the install strategy are stored in a
ConfigMap, this commit helps reducing the size of the
ConfigMap.

This is particularly necessary for the apiextensions/v1
migration as moving forward there will be multiple validation
schema per CRD which greatly increases the size of the ConfigMap
making it exceed the 1MiB limit.

To quickly inspect the ConfigMap:

kubectl -n kubevirt get cm -o json NAME | jq .data.manifests | tr -d '"' | base64 -d - | gunzip -

Signed-off-by: Antonio Cardace <acardace@redhat.com>